+++ app/src/main/java/com/shootr/mobile/ui/activities/MainTabbedActivity.java
@@ -4,25 +4,29 @@ import android.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
-import android.graphics.Color;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.annotation.IdRes;
 import android.support.v4.app.Fragment;
 import android.view.View;
-import android.view.animation.AccelerateDecelerateInterpolator;
+import android.widget.ImageView;
+import android.widget.LinearLayout;
+import android.widget.TextView;
 import butterknife.BindString;
+import butterknife.BindView;
 import butterknife.ButterKnife;
-import com.eftimoff.androidplayer.Player;
-import com.eftimoff.androidplayer.actions.property.PropertyAction;
+import butterknife.OnClick;
+import com.amulyakhare.textdrawable.TextDrawable;
 import com.roughike.bottombar.BottomBar;
-import com.roughike.bottombar.BottomBarBadge;
-import com.roughike.bottombar.OnMenuTabClickListener;
+import com.roughike.bottombar.BottomBarTab;
+import com.roughike.bottombar.OnTabReselectListener;
+import com.roughike.bottombar.OnTabSelectListener;
 import com.shootr.mobile.R;
 import com.shootr.mobile.ui.ToolbarDecorator;
 import com.shootr.mobile.ui.fragments.ActivityTimelineContainerFragment;
 import com.shootr.mobile.ui.fragments.FavoritesFragment;
 import com.shootr.mobile.ui.fragments.StreamsListFragment;
+import com.shootr.mobile.ui.model.StreamModel;
 import com.shootr.mobile.ui.model.UserModel;
 import com.shootr.mobile.ui.presenter.MainScreenPresenter;
 import com.shootr.mobile.ui.views.MainScreenView;
@@ -30,6 +34,9 @@ import com.shootr.mobile.util.CrashReportTool;
 import com.shootr.mobile.util.DeeplinkingNavigator;
 import com.shootr.mobile.util.DefaultTabUtils;
 import com.shootr.mobile.util.FeedbackMessage;
+import com.shootr.mobile.util.ImageLoader;
+import com.shootr.mobile.util.InitialsLoader;
+import de.hdodenhof.circleimageview.CircleImageView;
 import javax.inject.Inject;
 
 public class MainTabbedActivity extends BaseToolbarDecoratedActivity implements MainScreenView {
@@ -40,15 +47,21 @@ public class MainTabbedActivity extends BaseToolbarDecoratedActivity implements 
   private static final String EXTRA_MULTIPLE_ACTIVITIES = "multiple_activities";
   private static final int ACTIVITY_FRAGMENT = 3;
   @BindString(R.string.update_shootr_version_url) String updateVersionUrl;
+  @BindView(R.id.bottomBar) BottomBar bottomBar;
+  @BindView(R.id.connect_controller) LinearLayout connectController;
+  @BindView(R.id.stream_title) TextView streamTitle;
+  @BindView(R.id.stream_image) CircleImageView streamImage;
+  @BindView(R.id.stream_image_without_image) ImageView streamImageWithoutPicture;
   @Inject MainScreenPresenter mainScreenPresenter;
   @Inject FeedbackMessage feedbackMessage;
   @Inject DeeplinkingNavigator deeplinkingNavigator;
   @Inject DefaultTabUtils defaultTabUtils;
   @Inject CrashReportTool crashReportTool;
+  @Inject ImageLoader imageLoader;
+  @Inject InitialsLoader initialsLoader;
 
   private ToolbarDecorator toolbarDecorator;
-  private BottomBar bottomBar;
-  private BottomBarBadge unreadActivities;
+  private BottomBarTab activitiesTab;
   private Fragment currentFragment;
 
   public static Intent getUpdateNeededIntent(Context context) {
@@ -73,26 +86,14 @@ public class MainTabbedActivity extends BaseToolbarDecoratedActivity implements 
     loadIntentData();
     handleUpdateVersion();
     handleMultipleActivitiesIntent();
-    setupAnimation();
-  }
-
-  private void setupAnimation() {
-    PropertyAction bottomBarAction = PropertyAction.newPropertyAction(bottomBar)
-        .interpolator(new AccelerateDecelerateInterpolator())
-        .translationY(ANIMATION_TRANSLATION)
-        .duration(ANIMATION_DURATION)
-        .build();
-    Player.init().animate(bottomBarAction).play();
   }
 
   private void setupBottomBar(Bundle savedInstanceState) {
-    bottomBar = BottomBar.attach(findViewById(R.id.container), savedInstanceState);
-    bottomBar.noNavBarGoodness();
-    bottomBar.noTopOffset();
     setupBottomMenu();
-    bottomBar.setOnMenuTabClickListener(new OnMenuTabClickListener() {
-      @Override public void onMenuTabSelected(@IdRes int menuItemId) {
-        switch (menuItemId) {
+    activitiesTab = bottomBar.getTabWithId(R.id.bottombar_activity);
+    bottomBar.setOnTabSelectListener(new OnTabSelectListener() {
+      @Override public void onTabSelected(@IdRes int tabId) {
+        switch (tabId) {
           case R.id.bottombar_streams:
             StreamsListFragment streamsListFragment = StreamsListFragment.newInstance();
             currentFragment = streamsListFragment;
@@ -113,14 +114,16 @@ public class MainTabbedActivity extends BaseToolbarDecoratedActivity implements 
                 ActivityTimelineContainerFragment.newInstance();
             currentFragment = activityTimelineFragment;
             switchTab(activityTimelineFragment);
+            activitiesTab.removeBadge();
             break;
           default:
             break;
         }
       }
-
-      @Override public void onMenuTabReSelected(@IdRes int menuItemId) {
-        scrollToTop(menuItemId);
+    });
+    bottomBar.setOnTabReselectListener(new OnTabReselectListener() {
+      @Override public void onTabReSelected(@IdRes int tabId) {
+        scrollToTop(tabId);
       }
     });
     loadIntentData();
@@ -130,9 +133,9 @@ public class MainTabbedActivity extends BaseToolbarDecoratedActivity implements 
   private void setupBottomMenu() {
     if (defaultTabUtils.getDefaultTabPosition(getSessionHandler().getCurrentUserId())
         == DefaultTabUtils.DEFAULT_POSITION) {
-      bottomBar.setItems(R.menu.bottombar_menu);
+      bottomBar.setItems(R.xml.bottombar_menu);
     } else {
-      bottomBar.setItems(R.menu.bottombar_menu_discover);
+      bottomBar.setItems(R.xml.bottombar_menu_discover);
     }
   }
 
@@ -229,24 +232,51 @@ public class MainTabbedActivity extends BaseToolbarDecoratedActivity implements 
   }
 
   @Override public void showActivityBadge(int count) {
-    if (currentFragment != null &&
-        !(currentFragment instanceof ActivityTimelineContainerFragment)) {
+    if (currentFragment != null
+        && !(currentFragment instanceof ActivityTimelineContainerFragment)) {
       showBadge(count);
     }
   }
 
-  private void showBadge(int count) {
-    if (unreadActivities == null) {
-      unreadActivities = bottomBar.makeBadgeForTabAt(ACTIVITY_FRAGMENT, Color.TRANSPARENT, count);
+  @Override public void showConnectController(StreamModel streamModel) {
+    connectController.setVisibility(View.VISIBLE);
+    streamTitle.setText(streamModel.getTitle());
+    setupStreamPicture(streamModel);
+  }
+
+  private void setupStreamPicture(StreamModel streamModel) {
+    if (streamModel.getPicture() == null) {
+      String initials = initialsLoader.getLetters(streamModel.getTitle());
+      int backgroundColor = initialsLoader.getColorForLetters(initials);
+      TextDrawable textDrawable = initialsLoader.getTextDrawable(initials, backgroundColor);
+      streamImageWithoutPicture.setImageDrawable(textDrawable);
+      streamImageWithoutPicture.setVisibility(View.VISIBLE);
+      streamImage.setVisibility(View.GONE);
     } else {
-      unreadActivities.setCount(count);
-      unreadActivities.show();
+      imageLoader.load(streamModel.getPicture(), streamImage);
+      streamImageWithoutPicture.setVisibility(View.GONE);
+      streamImage.setVisibility(View.VISIBLE);
+    }
+  }
+
+  @Override public void hideConnectController() {
+    connectController.setVisibility(View.GONE);
+  }
+
+  @Override public void goToTimeline(StreamModel streamModel) {
+    startActivity(
+        StreamTimelineActivity.newIntent(this, streamModel.getIdStream(), streamModel.getTitle(),
+            streamModel.getAuthorId()));
+  }
+
+  private void showBadge(int count) {
+    if (activitiesTab != null) {
+      activitiesTab.setBadgeCount(count);
     }
   }
 
   @Override protected void onSaveInstanceState(Bundle outState) {
     super.onSaveInstanceState(outState);
-    bottomBar.onSaveInstanceState(outState);
   }
 
   private void setToolbarClickListener(final UserModel userModel) {
@@ -257,5 +287,13 @@ public class MainTabbedActivity extends BaseToolbarDecoratedActivity implements 
         overridePendingTransition(R.anim.slide_in_up, R.anim.stay);
       }
     });
+  }
+
+  @OnClick(R.id.close_button) public void onCloseClick() {
+    mainScreenPresenter.unwatchStream();
+  }
+
+  @OnClick(R.id.stream_info_container) public void onConnectControllerClick() {
+    mainScreenPresenter.onControllerClick();
   }
 }