+++ app/src/main/java/com/shootr/mobile/ui/activities/ProfileActivity.java
@@ -14,6 +14,7 @@ import android.os.Handler;
 import android.provider.MediaStore;
 import android.support.annotation.NonNull;
 import android.support.design.widget.CollapsingToolbarLayout;
+import android.support.design.widget.CoordinatorLayout;
 import android.support.design.widget.FloatingActionButton;
 import android.support.v4.content.FileProvider;
 import android.support.v7.app.ActionBar;
@@ -30,6 +31,7 @@ import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
 import com.cocosw.bottomsheet.BottomSheet;
+import com.github.clans.fab.FloatingActionMenu;
 import com.shootr.mobile.BuildConfig;
 import com.shootr.mobile.R;
 import com.shootr.mobile.domain.repository.SessionRepository;
@@ -116,6 +118,8 @@ public class ProfileActivity extends BaseActivity
 
   @BindView(R.id.profile_user_verified) ImageView userVerified;
   @BindView(R.id.mutuals_container) View mutualsContainer;
+  @BindView(R.id.profile_container) CoordinatorLayout profileContainer;
+  @BindView(R.id.fab_menu) FloatingActionMenu floatingMenu;
 
   @BindString(R.string.report_base_url) String reportBaseUrl;
   @BindString(R.string.analytics_screen_me) String analyticsScreenMe;
@@ -162,6 +166,8 @@ public class ProfileActivity extends BaseActivity
 
   private String idUser;
   private String username;
+
+  private int mScrollOffset = 4;
 
   public static Intent getIntent(Context context, String idUser) {
     Intent i = new Intent(context, ProfileActivity.class);
@@ -250,6 +256,8 @@ public class ProfileActivity extends BaseActivity
         reportShotPresenter.onShotLongPressed(shot);
       }
     });
+
+    floatingMenu.setClosedOnTouchOutside(true);
   }
 
   private void sendAnalytics(ShotModel shot) {
@@ -297,6 +305,18 @@ public class ProfileActivity extends BaseActivity
 
   @Override public void goToChannelTimeline(String idTargetUser) {
     startActivity(PrivateMessageTimelineActivity.newIntent(this, idTargetUser));
+  }
+
+  @Override public void hideFollowButton() {
+    followButton.setVisibility(View.GONE);
+  }
+
+  @Override public void hideEditMenu() {
+    floatingMenu.setVisibility(View.GONE);
+  }
+
+  @Override public void showEditMenu() {
+    floatingMenu.setVisibility(View.VISIBLE);
   }
 
   @Override protected void initializePresenter() {
@@ -1197,5 +1217,21 @@ public class ProfileActivity extends BaseActivity
 
   @OnClick(R.id.channel_button) public void onChannelClick() {
     profilePresenter.onChannelClick();
+  }
+
+  @OnClick(R.id.fab_edit_photo) public void onFabEditClick() {
+    closeFabMenu();
+    profilePresenter.avatarClicked();
+  }
+
+  @OnClick(R.id.fab_new_stream) public void onAddStream() {
+    closeFabMenu();
+    startActivityForResult(new Intent(this, NewStreamActivity.class), REQUEST_NEW_STREAM);
+  }
+
+  private void closeFabMenu() {
+    if (floatingMenu != null) {
+      floatingMenu.close(true);
+    }
   }
 }