+++ app/src/main/java/com/shootr/mobile/ui/adapters/listeners/OnOpenShotMenuListener.java
@@ -1,0 +1,8 @@
+package com.shootr.mobile.ui.adapters.listeners;
+
+import com.shootr.mobile.ui.model.ShotModel;
+
+public interface OnOpenShotMenuListener {
+
+    void openMenu(ShotModel shot);
+}