+++ app/src/main/java/com/shootr/mobile/ui/presenter/ProfilePresenter.java
@@ -142,8 +142,10 @@ public class ProfilePresenter implements Presenter {
     if (isCurrentUser) {
       profileView.showFriendsButton();
       profileView.hideChannelButton();
+      profileView.showEditMenu();
     } else {
       profileView.hideFriendsButton();
+      profileView.hideEditMenu();
     }
 
   }