+++ app/src/main/java/com/shootr/mobile/ui/views/MainScreenView.java
@@ -1,5 +1,6 @@
 package com.shootr.mobile.ui.views;
 
+import com.shootr.mobile.ui.model.StreamModel;
 import com.shootr.mobile.ui.model.UserModel;
 
 public interface MainScreenView {
@@ -7,4 +8,10 @@ public interface MainScreenView {
     void setUserData(UserModel userModel);
 
     void showActivityBadge(int count);
+
+    void showConnectController(StreamModel streamModel);
+
+    void hideConnectController();
+
+    void goToTimeline(StreamModel streamModel);
 }